backEnd/requirements.txt
Django==4.0.4
django-filter==22.1
django-cors-headers==3.11.0
django-rest-framework==0.1.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
drf-spectacular==0.22.0
environs==9.5.0
Pillow==9.1.0
python-dotenv==0.20.0
psycopg2-binary==2.9.5

gunicorn==20.1.0